+++ app/src/main/java/com/example/taskyapplication/agenda/di/DatabaseModule.kt
@@ -4,8 +4,8 @@ import android.content.Context
 import androidx.room.Room
 import com.example.taskyapplication.agenda.data.db.AgendaDatabase
 import com.example.taskyapplication.agenda.items.event.data.db.EventDao
-import com.example.taskyapplication.agenda.items.event.domain.EventLocalDataSource
-import com.example.taskyapplication.agenda.items.event.domain.EventLocalDataSourceImpl
+import com.example.taskyapplication.agenda.items.event.data.EventLocalDataSource
+import com.example.taskyapplication.agenda.items.event.data.EventLocalDataSourceImpl
 import com.example.taskyapplication.agenda.items.reminder.data.db.ReminderDao
 import com.example.taskyapplication.agenda.items.reminder.domain.ReminderLocalDataSource
 import com.example.taskyapplication.agenda.items.reminder.domain.ReminderLocalDataSourceImpl
