+++ app/src/main/java/com/example/taskyapplication/agenda/items/event/data/EventRemoteDataSource.kt
@@ -1,16 +1,6 @@
-package com.example.taskyapplication.agenda.items.event.domain
+package com.example.taskyapplication.agenda.items.event.data
 
-import com.example.taskyapplication.agenda.items.event.data.CreateEventNetworkModel
-import com.example.taskyapplication.agenda.items.event.data.CreatedEventResponse
-import com.example.taskyapplication.agenda.items.event.data.FetchedEventResponse
-import com.example.taskyapplication.agenda.items.event.data.GetAttendeeResponse
-import com.example.taskyapplication.agenda.items.event.data.UpdateEventNetworkModel
-import com.example.taskyapplication.agenda.items.event.data.UpdatedEventResponse
 import com.example.taskyapplication.agenda.items.event.network.EventApiService
-import com.example.taskyapplication.domain.utils.DataError
-import com.example.taskyapplication.domain.utils.EmptyResult
-import com.example.taskyapplication.domain.utils.Result
-import com.example.taskyapplication.domain.utils.safeApiCall
 import okhttp3.MultipartBody
 import retrofit2.Response
 import javax.inject.Inject
