+++ app/src/main/java/com/example/taskyapplication/agenda/items/event/domain/EventOfflineFirstRepository.kt
@@ -4,6 +4,8 @@ import android.content.Context
 import android.net.Uri
 import android.util.Log
 import com.example.taskyapplication.agenda.items.event.data.CreateEventNetworkModel
+import com.example.taskyapplication.agenda.items.event.data.EventLocalDataSource
+import com.example.taskyapplication.agenda.items.event.data.EventRemoteDataSource
 import com.example.taskyapplication.agenda.items.event.data.GetAttendeeResponse
 import com.example.taskyapplication.agenda.items.event.data.UpdateEventNetworkModel
 import com.example.taskyapplication.agenda.items.event.data.db.AttendeeEntity
