+++ app/src/main/java/com/example/taskyapplication/agenda/items/event/screens/EventDateTimeEdit.kt
@@ -12,7 +12,7 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
-import androidx.compose.material.Icon
+import androidx.compose.material3.Icon
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.outlined.CheckCircle
 import androidx.compose.material3.ExperimentalMaterial3Api
@@ -229,7 +229,7 @@ fun EventDateTimeScreen(
                 )
             },
             mainContent = {
-                var showDeleteBottomSheet by remember { mutableStateOf(false) }
+                var showDeleteBottomSheet by rememberSaveable { mutableStateOf(false) }
                 Box(
                     modifier = Modifier.fillMaxSize()
                 ) {
