+++ app/src/main/java/com/example/taskyapplication/auth/data/EmailValidator.kt
@@ -5,7 +5,7 @@ import com.example.taskyapplication.auth.domain.PatternValidator
 
 object EmailPatternValidator: PatternValidator {
 
-    override fun matches(email: String): Boolean {
-        return Patterns.EMAIL_ADDRESS.matcher(email).matches()
+    override fun matches(value: String): Boolean {
+        return Patterns.EMAIL_ADDRESS.matcher(value).matches()
     }
 }
